fraud-risk-pipeline/requirements.txt
duckdb
pandas
python-dotenv
requests
scikit-learn
fastapi
uvicorn
joblib
pytest